requirements.txt
tensorcircuit-ng>=1.1.0
qiskit==1.4.0
qiskit-algorithms==0.3.1
jax==0.4.37
jaxlib==0.4.36
joblib==1.4.2
optax
cotengra
tqdm